Back to the futures 3! (It's better than 2, I promise!)
Instead of having multiple threads polling Redis, now there's one that polls all the queues and delegates work to the agent thread pool with futures. Also, promises are now used in order to manage the number of workers (and to minimize polling). The polling thread will block (by dereferencing a promise) until a work is available, and then begin polling with that worker.